requirements.txt
toonify
toons
pydantic
tiktoken
openai
python-dotenv
scipy
